MIME batching refactorting ported from 6.x with 13 tests failinig
