+++ src/Microsoft.OData.Core/Microsoft.OData.Core.csproj
@@ -47,6 +47,9 @@
     <Compile Include="ContainerBuilderExtensions.cs" />
     <Compile Include="EdmExtensionMethods.cs" />
     <Compile Include="Evaluation\ODataConventionalResourceMetadataBuilder.cs" />
+    <Compile Include="MultipartMixed\ODataMultipartMixedBatchReader.cs" />
+    <Compile Include="MultipartMixed\ODataMultipartMixedBatchReaderStream.cs" />
+    <Compile Include="MultipartMixed\ODataMultipartMixedBatchWriter.cs" />
     <Compile Include="UnknownEntitySet.cs" />
     <Compile Include="IContainerProvider.cs" />
     <Compile Include="IDuplicatePropertyNameChecker.cs" />
