+++ src/Microsoft.OData.Core/ODataBatchReader.cs
@@ -17,24 +17,23 @@ namespace Microsoft.OData
     #endregion Namespaces
 
     /// <summary>
-    /// Class for reading OData batch messages; also verifies the proper sequence of read calls on the reader.
+    /// Abstract Class for reading OData batch messages; also verifies the proper sequence of read calls on the reader.
     /// </summary>
-    public sealed class ODataBatchReader : IODataBatchOperationListener
+    public abstract class ODataBatchReader : IODataBatchOperationListener
     {
         /// <summary>The input context to read the content from.</summary>
-        private readonly ODataRawInputContext inputContext;
+        private  ODataInputContext inputContext;
 
-        /// <summary>The batch stream used by the batch reader to devide a batch payload into parts.</summary>
-        private readonly ODataBatchReaderStream batchStream;
 
         /// <summary>True if the writer was created for synchronous operation; false for asynchronous.</summary>
         private readonly bool synchronous;
 
-        /// <summary>The batch-specific URL converter that stores the content IDs found in a changeset and supports resolving cross-referencing URLs.</summary>
-        private readonly ODataBatchPayloadUriConverter payloadUriConverter;
-
         /// <summary>The dependency injection container to get related services.</summary>
-        private readonly IServiceProvider container;
+        internal readonly IServiceProvider container;
+
+        /// <summary>The batch-specific URL converter that stores the content IDs found in a changeset and supports resolving cross-referencing URLs.</summary>
+        internal readonly ODataBatchPayloadUriConverter payloadUriConverter;
+
 
         /// <summary>The current state of the batch reader.</summary>
         private ODataBatchReaderState batchReaderState;
@@ -58,7 +57,7 @@ namespace Microsoft.OData
         /// <summary>
         /// Internal switch for whether we support reading Content-ID header appear in HTTP head instead of ChangeSet head.
         /// </summary>
-        private bool allowLegacyContentIdBehaviour;
+        private bool allowLegacyContentIdBehavior;
 
         /// <summary>
         /// Constructor.
@@ -67,23 +66,21 @@ namespace Microsoft.OData
         /// <param name="batchBoundary">The boundary string for the batch structure itself.</param>
         /// <param name="batchEncoding">The encoding to use to read from the batch stream.</param>
         /// <param name="synchronous">true if the reader is created for synchronous operation; false for asynchronous.</param>
-        internal ODataBatchReader(ODataRawInputContext inputContext, string batchBoundary, Encoding batchEncoding, bool synchronous)
+        protected ODataBatchReader(ODataInputContext inputContext, bool synchronous)
         {
             Debug.Assert(inputContext != null, "inputContext != null");
-            Debug.Assert(!string.IsNullOrEmpty(batchBoundary), "!string.IsNullOrEmpty(batchBoundary)");
 
             this.inputContext = inputContext;
             this.container = inputContext.Container;
             this.synchronous = synchronous;
             this.payloadUriConverter = new ODataBatchPayloadUriConverter(inputContext.PayloadUriConverter);
-            this.batchStream = new ODataBatchReaderStream(inputContext, batchBoundary, batchEncoding);
-            this.allowLegacyContentIdBehaviour = true;
+            this.allowLegacyContentIdBehavior = true;
         }
 
         /// <summary>
         /// An enumeration to track the state of a batch operation.
         /// </summary>
-        private enum OperationState
+        protected enum OperationState
         {
             /// <summary>No action has been performed on the operation.</summary>
             None,
@@ -98,6 +95,41 @@ namespace Microsoft.OData
             StreamDisposed,
         }
 
+        /// <summary>
+        /// Reader's input context.
+        /// </summary>
+        protected ODataInputContext InputContext
+        {
+            get { return this.inputContext; }
+        }
+
+        /// <summary>
+        /// Previously cache contentId that should be applied to the next message read.
+        /// </summary>
+        protected string ContentIdToAddOnNextRead
+        {
+            get { return this.contentIdToAddOnNextRead; }
+            set { this.contentIdToAddOnNextRead = value; }
+        }
+
+        /// <summary>
+        /// The reader's Operation state
+        /// </summary>
+        protected OperationState ReaderOperationState
+        {
+            get { return this.operationState; }
+            set { this.operationState = value; }
+        }
+
+        /// <summary>
+        /// Boolean flag indicating whether legacy content Id behavior is used.
+        /// </summary>
+        protected bool AllowLegacyContentIdBehavior
+        {
+            get { return this.allowLegacyContentIdBehavior; }
+        }
+
+
         /// <summary>Gets the current state of the batch reader.</summary>
         /// <returns>The current state of the batch reader.</returns>
         public ODataBatchReaderState State
@@ -108,7 +140,7 @@ namespace Microsoft.OData
                 return this.batchReaderState;
             }
 
-            private set
+            set
             {
                 this.batchReaderState = value;
             }
@@ -204,49 +236,57 @@ namespace Microsoft.OData
         }
 
         /// <summary>
-        /// Returns the next state of the batch reader after an end boundary has been found.
+        /// Increases the size of the current batch message; throws if the allowed limit is exceeded.
         /// </summary>
-        /// <returns>The next state of the batch reader.</returns>
-        private ODataBatchReaderState GetEndBoundaryState()
+        protected void IncreaseBatchSize()
         {
-            switch (this.batchReaderState)
+            this.currentBatchSize++;
+
+            if (this.currentBatchSize > this.inputContext.MessageReaderSettings.MessageQuotas.MaxPartsPerBatch)
             {
-                case ODataBatchReaderState.Initial:
-                    // If we find an end boundary when in state 'Initial' it means that we
-                    // have an empty batch. The next state will be 'Completed'.
-                    return ODataBatchReaderState.Completed;
-
-                case ODataBatchReaderState.Operation:
-                    // If we find an end boundary in state 'Operation' we have finished
-                    // processing an operation and found the end boundary of either the
-                    // current changeset or the batch.
-                    return this.batchStream.ChangeSetBoundary == null
-                        ? ODataBatchReaderState.Completed
-                        : ODataBatchReaderState.ChangesetEnd;
-
-                case ODataBatchReaderState.ChangesetStart:
-                    // If we find an end boundary when in state 'ChangeSetStart' it means that
-                    // we have an empty changeset. The next state will be 'ChangeSetEnd'
-                    return ODataBatchReaderState.ChangesetEnd;
-
-                case ODataBatchReaderState.ChangesetEnd:
-                    // If we are at the end of a changeset and find an end boundary
-                    // we reached the end of the batch
-                    return ODataBatchReaderState.Completed;
-
-                case ODataBatchReaderState.Completed:
-                    // We should never get here when in Completed state.
-                    throw new ODataException(Strings.General_InternalError(InternalErrorCodes.ODataBatchReader_GetEndBoundary_Completed));
-
-                case ODataBatchReaderState.Exception:
-                    // We should never get here when in Exception state.
-                    throw new ODataException(Strings.General_InternalError(InternalErrorCodes.ODataBatchReader_GetEndBoundary_Exception));
-
-                default:
-                    // Invalid enum value
-                    throw new ODataException(Strings.General_InternalError(InternalErrorCodes.ODataBatchReader_GetEndBoundary_UnknownValue));
+                throw new ODataException(Strings.ODataBatchReader_MaxBatchSizeExceeded(this.inputContext.MessageReaderSettings.MessageQuotas.MaxPartsPerBatch));
             }
         }
+
+        /// <summary>
+        /// Increases the size of the current change set; throws if the allowed limit is exceeded.
+        /// </summary>
+        protected void IncreaseChangesetSize()
+        {
+            this.currentChangeSetSize++;
+
+            if (this.currentChangeSetSize > this.inputContext.MessageReaderSettings.MessageQuotas.MaxOperationsPerChangeset)
+            {
+                throw new ODataException(Strings.ODataBatchReader_MaxChangeSetSizeExceeded(this.inputContext.MessageReaderSettings.MessageQuotas.MaxOperationsPerChangeset));
+            }
+        }
+
+        /// <summary>
+        /// Resets the size of the current change set to 0.
+        /// </summary>
+        protected void ResetChangesetSize()
+        {
+            this.currentChangeSetSize = 0;
+        }
+        /// <summary>
+        /// Continues reading from the batch message payload.
+        /// </summary>
+        /// <returns>true if more items were read; otherwise false.</returns>
+        protected abstract bool ReadImplementation();
+        /// <summary>
+        /// Returns the cached <see cref="ODataBatchOperationRequestMessage"/> for reading the content of an operation
+        /// in a batch request.
+        /// </summary>
+        /// <returns>The message that can be used to read the content of the batch request operation from.</returns>
+
+        protected abstract ODataBatchOperationRequestMessage CreateOperationRequestMessageImplementation();
+
+        /// <summary>
+        /// Returns the cached <see cref="ODataBatchOperationResponseMessage"/> for reading the content of a
+        /// batch response.
+        /// </summary>
+        /// <returns>The message that can be used to read the content of the batch response from.</returns>
+        protected abstract ODataBatchOperationResponseMessage CreateOperationResponseMessageImplementation();
 
         /// <summary>
         /// Reads the next part from the batch message payload.
@@ -262,6 +302,7 @@ namespace Microsoft.OData
         /// Asynchronously reads the next part from the batch message payload.
         /// </summary>
         /// <returns>A task that when completed indicates whether more information was read.</returns>
+        [SuppressMessage("Microsoft.MSInternal", "CA908:AvoidTypesThatRequireJitCompilationInPrecompiledAssemblies", Justification = "API design calls for a bool being returned from the task here.")]
         private Task<bool> ReadAsynchronously()
         {
             // We are reading from the fully buffered read stream here; thus it is ok
@@ -270,374 +311,6 @@ namespace Microsoft.OData
             return TaskUtils.GetTaskForSynchronousOperation<bool>(this.ReadImplementation);
         }
 #endif
-
-        /// <summary>
-        /// Continues reading from the batch message payload.
-        /// </summary>
-        /// <returns>true if more items were read; otherwise false.</returns>
-        private bool ReadImplementation()
-        {
-            Debug.Assert(this.operationState != OperationState.StreamRequested, "Should have verified that no operation stream is still active.");
-
-            switch (this.State)
-            {
-                case ODataBatchReaderState.Initial:
-                    // The stream should be positioned at the beginning of the batch content,
-                    // that is before the first boundary (or the preamble if there is one).
-                    this.batchReaderState = this.SkipToNextPartAndReadHeaders();
-                    break;
-
-                case ODataBatchReaderState.Operation:
-                    // When reaching this state we already read the MIME headers of the operation.
-                    // Clients MUST call CreateOperationRequestMessage
-                    // or CreateOperationResponseMessage to read at least the headers of the operation.
-                    // This is important since we need to read the ContentId header (if present) and
-                    // add it to the URL resolver.
-                    if (this.operationState == OperationState.None)
-                    {
-                        // No message was created; fail
-                        throw new ODataException(Strings.ODataBatchReader_NoMessageWasCreatedForOperation);
-                    }
-
-                    // Reset the operation state; the operation state only
-                    // tracks the state of a batch operation while in state Operation.
-                    this.operationState = OperationState.None;
-
-                    // Also add a pending ContentId header to the URL resolver now. We ensured above
-                    // that a message has been created for this operation and thus the headers (incl.
-                    // a potential content ID header) have been read.
-                    if (this.contentIdToAddOnNextRead != null)
-                    {
-                        this.payloadUriConverter.AddContentId(this.contentIdToAddOnNextRead);
-                        this.contentIdToAddOnNextRead = null;
-                    }
-
-                    // When we are done with an operation, we have to skip ahead to the next part
-                    // when Read is called again. Note that if the operation stream was never requested
-                    // and the content of the operation has not been read, we'll skip it here.
-                    Debug.Assert(this.operationState == OperationState.None, "Operation state must be 'None' at the end of the operation.");
-                    this.batchReaderState = this.SkipToNextPartAndReadHeaders();
-                    break;
-
-                case ODataBatchReaderState.ChangesetStart:
-                    // When at the start of a changeset, skip ahead to the first operation in the
-                    // changeset (or the end boundary of the changeset).
-                    Debug.Assert(this.batchStream.ChangeSetBoundary != null, "Changeset boundary must have been set by now.");
-                    this.batchReaderState = this.SkipToNextPartAndReadHeaders();
-                    break;
-
-                case ODataBatchReaderState.ChangesetEnd:
-                    // When at the end of a changeset, reset the changeset boundary and the
-                    // changeset size and then skip to the next part.
-                    this.ResetChangeSetSize();
-                    this.batchStream.ResetChangeSetBoundary();
-                    this.batchReaderState = this.SkipToNextPartAndReadHeaders();
-                    break;
-
-                case ODataBatchReaderState.Exception:    // fall through
-                case ODataBatchReaderState.Completed:
-                    Debug.Assert(false, "Should have checked in VerifyCanRead that we are not in one of these states.");
-                    throw new ODataException(Strings.General_InternalError(InternalErrorCodes.ODataBatchReader_ReadImplementation));
-
-                default:
-                    Debug.Assert(false, "Unsupported reader state " + this.State + " detected.");
-                    throw new ODataException(Strings.General_InternalError(InternalErrorCodes.ODataBatchReader_ReadImplementation));
-            }
-
-            return this.batchReaderState != ODataBatchReaderState.Completed && this.batchReaderState != ODataBatchReaderState.Exception;
-        }
-
-        /// <summary>
-        /// Skips all data in the stream until the next part is detected; then reads the part's request/response line and headers.
-        /// </summary>
-        /// <returns>The next state of the batch reader after skipping to the next part and reading the part's beginning.</returns>
-        private ODataBatchReaderState SkipToNextPartAndReadHeaders()
-        {
-            bool isEndBoundary, isParentBoundary;
-            bool foundBoundary = this.batchStream.SkipToBoundary(out isEndBoundary, out isParentBoundary);
-
-            if (!foundBoundary)
-            {
-                // We did not find the expected boundary at all in the payload;
-                // we are done reading. Depending on where we are report changeset end or completed state
-                if (this.batchStream.ChangeSetBoundary == null)
-                {
-                    return ODataBatchReaderState.Completed;
-                }
-                else
-                {
-                    return ODataBatchReaderState.ChangesetEnd;
-                }
-            }
-
-            ODataBatchReaderState nextState;
-            if (isEndBoundary || isParentBoundary)
-            {
-                // We detected an end boundary or detected that the end boundary is missing
-                // because we found a parent boundary
-                nextState = this.GetEndBoundaryState();
-
-                if (nextState == ODataBatchReaderState.ChangesetEnd)
-                {
-                    // Reset the URL resolver at the end of a changeset; Content IDs are
-                    // unique within a given changeset.
-                    this.payloadUriConverter.Reset();
-                }
-            }
-            else
-            {
-                bool currentlyInChangeSet = this.batchStream.ChangeSetBoundary != null;
-                string contentId;
-
-                // If we did not find an end boundary, we found another part
-                bool isChangeSetPart = this.batchStream.ProcessPartHeader(out contentId);
-
-                // Compute the next reader state
-                if (currentlyInChangeSet)
-                {
-                    Debug.Assert(!isChangeSetPart, "Should have validated that nested changesets are not allowed.");
-                    nextState = ODataBatchReaderState.Operation;
-                    this.IncreaseChangeSetSize();
-                }
-                else
-                {
-                    // We are at the top level (not inside a changeset)
-                    nextState = isChangeSetPart
-                        ? ODataBatchReaderState.ChangesetStart
-                        : ODataBatchReaderState.Operation;
-                    this.IncreaseBatchSize();
-                }
-
-                if (!isChangeSetPart)
-                {
-                    // Add a potential Content-ID header to the URL resolver so that it will be available
-                    // to subsequent operations.
-                    Debug.Assert(this.contentIdToAddOnNextRead == null, "Must not have a content ID to be added to a part.");
-
-                    if (contentId != null && this.payloadUriConverter.ContainsContentId(contentId))
-                    {
-                        throw new ODataException(Strings.ODataBatchReader_DuplicateContentIDsNotAllowed(contentId));
-                    }
-
-                    this.contentIdToAddOnNextRead = contentId;
-                }
-            }
-
-            return nextState;
-        }
-
-        /// <summary>
-        /// Returns the cached <see cref="ODataBatchOperationRequestMessage"/> for reading the content of an operation
-        /// in a batch request.
-        /// </summary>
-        /// <returns>The message that can be used to read the content of the batch request operation from.</returns>
-        private ODataBatchOperationRequestMessage CreateOperationRequestMessageImplementation()
-        {
-            this.operationState = OperationState.MessageCreated;
-
-            string requestLine = this.batchStream.ReadFirstNonEmptyLine();
-
-            string httpMethod;
-            Uri requestUri;
-            this.ParseRequestLine(requestLine, out httpMethod, out requestUri);
-
-            // Read all headers and create the request message
-            ODataBatchOperationHeaders headers = this.batchStream.ReadHeaders();
-
-            if (this.batchStream.ChangeSetBoundary != null)
-            {
-                if (this.allowLegacyContentIdBehaviour)
-                {
-                    // Add a potential Content-ID header to the URL resolver so that it will be available
-                    // to subsequent operations.
-                    string contentId;
-                    if (this.contentIdToAddOnNextRead == null && headers.TryGetValue(ODataConstants.ContentIdHeader, out contentId))
-                    {
-                        if (contentId != null && this.payloadUriConverter.ContainsContentId(contentId))
-                        {
-                            throw new ODataException(Strings.ODataBatchReader_DuplicateContentIDsNotAllowed(contentId));
-                        }
-
-                        this.contentIdToAddOnNextRead = contentId;
-                    }
-                }
-
-                if (this.contentIdToAddOnNextRead == null)
-                {
-                    throw new ODataException(Strings.ODataBatchOperationHeaderDictionary_KeyNotFound(ODataConstants.ContentIdHeader));
-                }
-            }
-
-            ODataBatchOperationRequestMessage requestMessage = ODataBatchOperationRequestMessage.CreateReadMessage(
-                this.batchStream,
-                httpMethod,
-                requestUri,
-                headers,
-                /*operationListener*/ this,
-                this.contentIdToAddOnNextRead,
-                this.payloadUriConverter,
-                this.container);
-
-            return requestMessage;
-        }
-
-        /// <summary>
-        /// Returns the cached <see cref="ODataBatchOperationRequestMessage"/> for reading the content of an operation
-        /// in a batch request.
-        /// </summary>
-        /// <returns>The message that can be used to read the content of the batch request operation from.</returns>
-        private ODataBatchOperationResponseMessage CreateOperationResponseMessageImplementation()
-        {
-            this.operationState = OperationState.MessageCreated;
-
-            string responseLine = this.batchStream.ReadFirstNonEmptyLine();
-
-            int statusCode = this.ParseResponseLine(responseLine);
-
-            // Read all headers and create the response message
-            ODataBatchOperationHeaders headers = this.batchStream.ReadHeaders();
-
-            if (this.batchStream.ChangeSetBoundary != null)
-            {
-                if (this.allowLegacyContentIdBehaviour)
-                {
-                    // Add a potential Content-ID header to the URL resolver so that it will be available
-                    // to subsequent operations.
-                    string contentId;
-                    if (this.contentIdToAddOnNextRead == null && headers.TryGetValue(ODataConstants.ContentIdHeader, out contentId))
-                    {
-                        if (contentId != null && this.payloadUriConverter.ContainsContentId(contentId))
-                        {
-                            throw new ODataException(Strings.ODataBatchReader_DuplicateContentIDsNotAllowed(contentId));
-                        }
-
-                        this.contentIdToAddOnNextRead = contentId;
-                    }
-                }
-            }
-
-            // In responses we don't need to use our batch URL resolver, since there are no cross referencing URLs
-            // so use the URL resolver from the batch message instead.
-            ODataBatchOperationResponseMessage responseMessage = ODataBatchOperationResponseMessage.CreateReadMessage(
-                this.batchStream,
-                statusCode,
-                headers,
-                this.contentIdToAddOnNextRead,
-                /*operationListener*/ this,
-                this.payloadUriConverter.BatchMessagePayloadUriConverter,
-                this.container);
-
-            //// NOTE: Content-IDs for cross referencing are only supported in request messages; in responses
-            ////       we allow a Content-ID header but don't process it (i.e., don't add the content ID to the URL resolver).
-
-            return responseMessage;
-        }
-
-        /// <summary>
-        /// Parses the request line of a batch operation request.
-        /// </summary>
-        /// <param name="requestLine">The request line as a string.</param>
-        /// <param name="httpMethod">The parsed HTTP method of the request.</param>
-        /// <param name="requestUri">The parsed <see cref="Uri"/> of the request.</param>
-        private void ParseRequestLine(string requestLine, out string httpMethod, out Uri requestUri)
-        {
-            Debug.Assert(!this.inputContext.ReadingResponse, "Must only be called for requests.");
-
-            // Batch Request: POST /Customers HTTP/1.1
-            // Since the uri can contain spaces, the only way to read the request url, is to
-            // check for first space character and last space character and anything between
-            // them.
-            int firstSpaceIndex = requestLine.IndexOf(' ');
-
-            // Check whether there are enough characters after the first space for the 2nd and 3rd segments
-            // (and a whitespace in between)
-            if (firstSpaceIndex <= 0 || requestLine.Length - 3 <= firstSpaceIndex)
-            {
-                // only 1 segment or empty first segment or not enough left for 2nd and 3rd segments
-                throw new ODataException(Strings.ODataBatchReaderStream_InvalidRequestLine(requestLine));
-            }
-
-            int lastSpaceIndex = requestLine.LastIndexOf(' ');
-            if (lastSpaceIndex < 0 || lastSpaceIndex - firstSpaceIndex - 1 <= 0 || requestLine.Length - 1 <= lastSpaceIndex)
-            {
-                // only 2 segments or empty 2nd or 3rd segments
-                // only 1 segment or empty first segment or not enough left for 2nd and 3rd segments
-                throw new ODataException(Strings.ODataBatchReaderStream_InvalidRequestLine(requestLine));
-            }
-
-            httpMethod = requestLine.Substring(0, firstSpaceIndex);               // Request - Http method
-            string uriSegment = requestLine.Substring(firstSpaceIndex + 1, lastSpaceIndex - firstSpaceIndex - 1);      // Request - Request uri
-            string httpVersionSegment = requestLine.Substring(lastSpaceIndex + 1);             // Request - Http version
-
-            // Validate HttpVersion
-            if (string.CompareOrdinal(ODataConstants.HttpVersionInBatching, httpVersionSegment) != 0)
-            {
-                throw new ODataException(Strings.ODataBatchReaderStream_InvalidHttpVersionSpecified(httpVersionSegment, ODataConstants.HttpVersionInBatching));
-            }
-
-            // NOTE: this method will throw if the method is not recognized.
-            HttpUtils.ValidateHttpMethod(httpMethod);
-
-            // Validate the HTTP method when reading a request
-            if (this.batchStream.ChangeSetBoundary != null)
-            {
-                // allow all methods except for GET
-                if (HttpUtils.IsQueryMethod(httpMethod))
-                {
-                    throw new ODataException(Strings.ODataBatch_InvalidHttpMethodForChangeSetRequest(httpMethod));
-                }
-            }
-
-            requestUri = new Uri(uriSegment, UriKind.RelativeOrAbsolute);
-            requestUri = ODataBatchUtils.CreateOperationRequestUri(requestUri, this.inputContext.MessageReaderSettings.BaseUri, this.payloadUriConverter);
-        }
-
-        /// <summary>
-        /// Parses the response line of a batch operation response.
-        /// </summary>
-        /// <param name="responseLine">The response line as a string.</param>
-        /// <returns>The parsed status code from the response line.</returns>
-        [SuppressMessage("Microsoft.Performance", "CA1822:MarkMembersAsStatic", Justification = "'this' is used when built in debug")]
-        private int ParseResponseLine(string responseLine)
-        {
-            Debug.Assert(this.inputContext.ReadingResponse, "Must only be called for responses.");
-
-            // Batch Response: HTTP/1.1 200 Ok
-            // Since the http status code strings have spaces in them, we cannot use the same
-            // logic. We need to check for the second space and anything after that is the error
-            // message.
-            int firstSpaceIndex = responseLine.IndexOf(' ');
-            if (firstSpaceIndex <= 0 || responseLine.Length - 3 <= firstSpaceIndex)
-            {
-                // only 1 segment or empty first segment or not enough left for 2nd and 3rd segments
-                throw new ODataException(Strings.ODataBatchReaderStream_InvalidResponseLine(responseLine));
-            }
-
-            int secondSpaceIndex = responseLine.IndexOf(' ', firstSpaceIndex + 1);
-            if (secondSpaceIndex < 0 || secondSpaceIndex - firstSpaceIndex - 1 <= 0 || responseLine.Length - 1 <= secondSpaceIndex)
-            {
-                // only 2 segments or empty 2nd or 3rd segments
-                // only 1 segment or empty first segment or not enough left for 2nd and 3rd segments
-                throw new ODataException(Strings.ODataBatchReaderStream_InvalidResponseLine(responseLine));
-            }
-
-            string httpVersionSegment = responseLine.Substring(0, firstSpaceIndex);
-            string statusCodeSegment = responseLine.Substring(firstSpaceIndex + 1, secondSpaceIndex - firstSpaceIndex - 1);
-
-            // Validate HttpVersion
-            if (string.CompareOrdinal(ODataConstants.HttpVersionInBatching, httpVersionSegment) != 0)
-            {
-                throw new ODataException(Strings.ODataBatchReaderStream_InvalidHttpVersionSpecified(httpVersionSegment, ODataConstants.HttpVersionInBatching));
-            }
-
-            int intResult;
-            if (!Int32.TryParse(statusCodeSegment, out intResult))
-            {
-                throw new ODataException(Strings.ODataBatchReaderStream_NonIntegerHttpStatusCode(statusCodeSegment));
-            }
-
-            return intResult;
-        }
 
         /// <summary>
         /// Verifies that calling CreateOperationRequestMessage if valid.
@@ -742,40 +415,6 @@ namespace Microsoft.OData
                 Debug.Assert(false, "Async calls are not allowed in this build.");
 #endif
             }
-        }
-
-        /// <summary>
-        /// Increases the size of the current batch message; throws if the allowed limit is exceeded.
-        /// </summary>
-        private void IncreaseBatchSize()
-        {
-            this.currentBatchSize++;
-
-            if (this.currentBatchSize > this.inputContext.MessageReaderSettings.MessageQuotas.MaxPartsPerBatch)
-            {
-                throw new ODataException(Strings.ODataBatchReader_MaxBatchSizeExceeded(this.inputContext.MessageReaderSettings.MessageQuotas.MaxPartsPerBatch));
-            }
-        }
-
-        /// <summary>
-        /// Increases the size of the current change set; throws if the allowed limit is exceeded.
-        /// </summary>
-        private void IncreaseChangeSetSize()
-        {
-            this.currentChangeSetSize++;
-
-            if (this.currentChangeSetSize > this.inputContext.MessageReaderSettings.MessageQuotas.MaxOperationsPerChangeset)
-            {
-                throw new ODataException(Strings.ODataBatchReader_MaxChangeSetSizeExceeded(this.inputContext.MessageReaderSettings.MessageQuotas.MaxOperationsPerChangeset));
-            }
-        }
-
-        /// <summary>
-        /// Resets the size of the current change set to 0.
-        /// </summary>
-        private void ResetChangeSetSize()
-        {
-            this.currentChangeSetSize = 0;
         }
 
         /// <summary>
